test manual contract verification
